refactor: re-use some code, add endpoint for single component
