+++ modules/analysis/README.md
@@ -1,7 +1,21 @@
 # Analysis Graph
 
-## Get a component
+## Get a root component
 
 ```bash
 http localhost:8080/api/v2/analysis/root-component/B
+```
+
+## Get a component
+
+With the name `B`:
+
+```bash
+http localhost:8080/api/v2/analysis/component/B
+```
+
+With the PURL ``:
+
+```bash
+http localhost:8080/api/v2/analysis/component/B
 ```
